Bug fix:   PNG export: indexed option doesn't work     (feature initially contributed by: Jun Inamori <jun@oop-reserch.com>);     (fix contributed by: x00000000@freenet.de);     (closes bug 43847).
General:
  Few whitespace fixes.

git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@907280 13f79535-47bb-0310-9956-ffa450edef68

+++ sources/org/apache/batik/apps/svgbrowser/JSVGViewerFrame.java
@@ -1764,10 +1764,7 @@ public class JSVGViewerFrame
 
             int choice = fileChooser.showSaveDialog(JSVGViewerFrame.this);
             if (choice == JFileChooser.APPROVE_OPTION) {
-
-            // Start: By Jun Inamori (jun@oop-reserch.com)
-            boolean isIndexed = PNGOptionPanel.showDialog(JSVGViewerFrame.this);
-            // End: By Jun Inamori (jun@oop-reserch.com)
+                boolean isIndexed = PNGOptionPanel.showDialog(JSVGViewerFrame.this);
 
                 final File f = fileChooser.getSelectedFile();
                 BufferedImage buffer = svgCanvas.getOffScreen();
@@ -1787,11 +1784,9 @@ public class JSVGViewerFrame
                     trans.addTranscodingHint(PNGTranscoder.KEY_FORCE_TRANSPARENT_WHITE,
                                              Boolean.TRUE );
 
-            // Start: By Jun Inamori
-            if(isIndexed){
-                trans.addTranscodingHint(PNGTranscoder.KEY_INDEXED,new Integer(256));
-            }
-            // End: By Jun Inamori
+                    if(isIndexed){
+                        trans.addTranscodingHint(PNGTranscoder.KEY_INDEXED, new Integer(8));
+                    }
 
                     final BufferedImage img = trans.createImage(w, h);
